Added DTO Files for Details and Resume Projet, Finished Up Update Method for Projet, Changed double types to decimal types

+++ PlantC.CitoyensEntreprise.DAL/Entities/Localisation.cs
@@ -3,7 +3,7 @@
         public int Id { get; set; }
         public string NomLocalite { get; set; }
         public short CodePostal { get; set; }
-        public double Longitude { get; set; }
-        public double Latitude { get; set; }
+        public decimal Longitude { get; set; }
+        public decimal Latitude { get; set; }
     }
 }
