+++ PlantC.CitoyensEntreprises.BLL/Models/MarqueurModel.cs
@@ -8,7 +8,7 @@ namespace PlantC.CitoyensEntreprises.BLL.Models {
     public class MarqueurModel {
         public int IdProjet { get; set; }
         public string Infrastructure { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
     }
 }
